(streamlit_app.py | dashboard_utils.py) - Adjusting dashboard code adding an utils lib to centralize certain codes
