+++ dashboard/streamlit_app.py
@@ -4,6 +4,8 @@ import pandas as pd
 import requests
 import plotly.graph_objects as go
 import numpy as np
+from dashboard_utils import (get_fear_greed_index, 
+                             get_bitcoin_data)
 
 
 def get_prediction():
@@ -16,7 +18,6 @@ def get_prediction():
         "date": latest_data.index[0].strftime("%Y-%m-%d"),
         "low": latest_data["Low"].iloc[0],
         "close": latest_data["Close"].iloc[0],
-        # Add other features here
         # "sentiment":
         # "neg_sentiment":
         # "close_ratio_2":
@@ -39,29 +40,29 @@ def get_prediction():
         return None, None
 
 
-def get_bitcoin_data():
-    btc = yf.Ticker("BTC-USD")
-    data = btc.history(period="1mo")
-    return get_last_7_days(data)
+# def get_bitcoin_data():
+#     btc = yf.Ticker("BTC-USD")
+#     data = btc.history(period="1mo")
+#     return get_last_7_days(data)
 
 
-def get_last_7_days(data):
-    end_date = data.index[-1]
-    start_date = end_date - pd.Timedelta(days=7)
-    return data.loc[start_date:end_date]
+# def get_last_7_days(data):
+#     end_date = data.index[-1]
+#     start_date = end_date - pd.Timedelta(days=7)
+#     return data.loc[start_date:end_date]
 
 
-def get_fear_greed_index():
-    url = "https://api.alternative.me/fng/?limit=7"
-    response = requests.get(url)
-    if response.status_code == 200:
-        data = response.json()["data"]
-        df = pd.DataFrame(data)
-        df["timestamp"] = pd.to_datetime(df["timestamp"].astype(float), unit="s")
-        df = df.sort_values("timestamp")
-        return df
-    else:
-        return None
+# def get_fear_greed_index():
+#     url = "https://api.alternative.me/fng/?limit=7"
+#     response = requests.get(url)
+#     if response.status_code == 200:
+#         data = response.json()["data"]
+#         df = pd.DataFrame(data)
+#         df["timestamp"] = pd.to_datetime(df["timestamp"].astype(float), unit="s")
+#         df = df.sort_values("timestamp")
+#         return df
+#     else:
+#         return None
 
 
 st.title("💰 Bitcoin Trend Prediction")
